ISA/requirements.txt
adal>=1.2.7
asgiref>=3.6.0
async-timeout>=4.0.2
azure-common>=1.1.28
azure-core>=1.26.3
azure-graphrbac>=0.61.1
azure-identity>=1.12.0
backports.functools-lru-cache>=1.6.4
certifi>=2023.7.22
cffi>=1.15.1
charset-normalizer>=3.0.1
click>=8.1.3
colorama>=0.4.6
cryptography>=41.0.3
distlib>=0.3.6
Django>=4.2.4
django-auth-adfs>=1.11.4
django-parler>=2.3
django-rosetta>=0.9.8
django-rq>=2.4.1
django-sslserver>=0.22
djangorestframework>=3.14.0
filelock>=3.9.0
gevent>=22.10.2
greenlet>=2.0.2
gunicorn>=20.1.0
idna>=3.4
isodate>=0.6.1
msal>=1.21.0
msal-extensions>=1.0.0
msgraph-core>=0.2.2
msrest>=0.7.1
msrestazure>=0.6.4
oauthlib>=3.2.2
platformdirs>=3.0.0
polib>=1.2.0
portalocker>=2.7.0
psycopg>=3.1.8
psycopg2>=2.9.5
psycopg2-binary>=2.9.5
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.21
PyJWT>=2.6.0
python-dateutil>=2.8.2
pytz>=2022.7.1
redis>=4.6.0
requests>=2.31.0
requests-oauthlib>=1.3.1
rq>=1.13.0
six>=1.16.0
six21>=0.1.3
solid-i18n>=1.4.2
sqlparse>=0.4.4
typing_extensions>=4.4.0
tzdata>=2022.7
urllib3>=1.26.14
wfastcgi>=3.0.0
zope.event>=4.6
zope.interface>=6.0
